Move remaining crypto provider functionality into SecureStorage.
